Version-05/04: backup code, have error is not shipper success, delivery not show

+++ ElogictisMobile/ElogictisMobile/Views/AddProductFormPage.xaml.cs
@@ -29,7 +29,7 @@ namespace ElogictisMobile.Views
             }    
             if (this.BindingContext is AddProductFormPageViewModel vm)
             {
-                vm.WeightChangedCommand.Execute(int.Parse(e.NewTextValue));
+                vm.WeightChangedCommand.Execute(double.Parse(e.NewTextValue));
             }
         }
